Add files via upload
Made this code while following steps from MIT lecture 
Stern-Brocot trees are named after Moritz Stern and Achille Brocot who independently discovered them. A Stern-Brocot tree is an infinite, complete binary-search tree whose nodes stand in a one-to-one correspondence with the positive rational numbers, and it is used to find the closest rational approximation to an arbitrary floating-point number. The following figure, taken from Wikipedia, illustrates a partial 4-level Stern-Brocot tree.
The rationals stored in each node are caculated using the following simple rule. The value stored in the left child of a given parent node equals the value in the parent plus the rational on the vertical line to the left of the child. However, the additions are not performed in ordinary rational arithmetic, but simply by adding the corresponding numerators and denominators. For example, the rationals on the leftmost branch of the tree under the root are: (1+0)/(1+1) = 1/2, (1+0)/(2+1) = 1/3, (1+0)/(3+1) = 1/4. Likewise, the left child of the 2/3 parent node contains the rational (2+1)/(3+2) = 3/5. In the case of right children of parent nodes, they contain the rational of the parent plus the rational on the vertical line to their right. For example, the rationals on the rightmost branch of the tree under the root are: (1+1)/(1+0) = 2/1, (2+1)/(1+0) = 3/1, (3+1)/(1+0) = 4/1. Similarly, the right child of the 3/2 parent node contains the rational (3+2)/(2+1) = 5/3. The results of all these additions are called mediants. In general, given two rational numbers p1/n1 and p2/n2, their mediant is (p1 + p2)/(n1 + n2).
The fundamental data structure for Stern-Brocot trees is defined by a class to implement its nodes. A node contains two integers _p and _q , corresponding to the numerator and denominator of a rational number, and references to its left and right children plus a reference to its parent. By definition, the parent of the root of the tree is null.